+++ src/py/flwr/cli/config_utils_test.py
@@ -23,8 +23,6 @@ from typing import Any, Optional
 import click
 import pytest
 
-from flwr.common.config import validate_config, validate_fields_in_config
-
 from .config_utils import (
     load,
     process_loaded_project_config,
@@ -161,184 +159,6 @@ def test_load_pyproject_toml_from_path(tmp_path: Path) -> None:
         assert config == expected_config
     finally:
         os.chdir(origin)
-
-
-def test_validate_pyproject_toml_fields_empty() -> None:
-    """Test that validate_pyproject_toml_fields fails correctly."""
-    # Prepare
-    config: dict[str, Any] = {}
-
-    # Execute
-    is_valid, errors, warnings = validate_fields_in_config(config)
-
-    # Assert
-    assert not is_valid
-    assert len(errors) == 2
-    assert len(warnings) == 0
-
-
-def test_validate_pyproject_toml_fields_no_flower() -> None:
-    """Test that validate_pyproject_toml_fields fails correctly."""
-    # Prepare
-    config = {
-        "project": {
-            "name": "fedgpt",
-            "version": "1.0.0",
-            "description": "",
-            "license": "",
-            "authors": [],
-        }
-    }
-
-    # Execute
-    is_valid, errors, warnings = validate_fields_in_config(config)
-
-    # Assert
-    assert not is_valid
-    assert len(errors) == 1
-    assert len(warnings) == 0
-
-
-def test_validate_pyproject_toml_fields_no_flower_components() -> None:
-    """Test that validate_pyproject_toml_fields fails correctly."""
-    # Prepare
-    config = {
-        "project": {
-            "name": "fedgpt",
-            "version": "1.0.0",
-            "description": "",
-            "license": "",
-            "authors": [],
-        },
-        "tool": {"flwr": {"app": {}}},
-    }
-
-    # Execute
-    is_valid, errors, warnings = validate_fields_in_config(config)
-
-    # Assert
-    assert not is_valid
-    assert len(errors) == 2
-    assert len(warnings) == 0
-
-
-def test_validate_pyproject_toml_fields_no_server_and_client_app() -> None:
-    """Test that validate_pyproject_toml_fields fails correctly."""
-    # Prepare
-    config = {
-        "project": {
-            "name": "fedgpt",
-            "version": "1.0.0",
-            "description": "",
-            "license": "",
-            "authors": [],
-        },
-        "tool": {"flwr": {"app": {"components": {}}}},
-    }
-
-    # Execute
-    is_valid, errors, warnings = validate_fields_in_config(config)
-
-    # Assert
-    assert not is_valid
-    assert len(errors) == 3
-    assert len(warnings) == 0
-
-
-def test_validate_pyproject_toml_fields() -> None:
-    """Test that validate_pyproject_toml_fields succeeds correctly."""
-    # Prepare
-    config = {
-        "project": {
-            "name": "fedgpt",
-            "version": "1.0.0",
-            "description": "",
-            "license": "",
-            "authors": [],
-        },
-        "tool": {
-            "flwr": {
-                "app": {
-                    "publisher": "flwrlabs",
-                    "components": {"serverapp": "", "clientapp": ""},
-                },
-            },
-        },
-    }
-
-    # Execute
-    is_valid, errors, warnings = validate_fields_in_config(config)
-
-    # Assert
-    assert is_valid
-    assert len(errors) == 0
-    assert len(warnings) == 0
-
-
-def test_validate_pyproject_toml() -> None:
-    """Test that validate_pyproject_toml succeeds correctly."""
-    # Prepare
-    config = {
-        "project": {
-            "name": "fedgpt",
-            "version": "1.0.0",
-            "description": "",
-            "license": "",
-            "authors": [],
-        },
-        "tool": {
-            "flwr": {
-                "app": {
-                    "publisher": "flwrlabs",
-                    "components": {
-                        "serverapp": "flwr.cli.run:run",
-                        "clientapp": "flwr.cli.run:run",
-                    },
-                },
-            },
-        },
-    }
-
-    # Execute
-    is_valid, errors, warnings = validate_config(config)
-
-    # Assert
-    assert is_valid
-    assert not errors
-    assert not warnings
-
-
-def test_validate_pyproject_toml_fail() -> None:
-    """Test that validate_pyproject_toml fails correctly."""
-    # Prepare
-    config = {
-        "project": {
-            "name": "fedgpt",
-            "version": "1.0.0",
-            "description": "",
-            "license": "",
-            "authors": [],
-        },
-        "tool": {
-            "flwr": {
-                "app": {
-                    "publisher": "flwrlabs",
-                    "components": {
-                        "serverapp": "flwr.cli.run:run",
-                        "clientapp": "flwr.cli.run:runa",
-                    },
-                },
-            },
-        },
-    }
-
-    # Execute
-    is_valid, errors, warnings = validate_config(config)
-
-    # Assert
-    assert not is_valid
-    assert len(errors) == 1
-    assert len(warnings) == 0
 
 
 def test_validate_project_config_fail() -> None:
